// JSSMarineWeb/untitled folder/src/app/api-middleware/model/cUAuth.ts
/**
 * Catchup APIs
 * The Catchup API document is intended strictly for internal use only.
 *
 * OpenAPI spec version: 1.0.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface CUAuth {
  /**
   * user id of the user.
   */
  user_id: string;
  /**
   * registration key of the user.
   */
  app_key: string;
}